py/requirements.txt
paho-mqtt==1.6.1
json5==0.9.10
